Move time_now resolution in Contracts
